+++ HISTORY.rst
@@ -12,6 +12,7 @@ End-User Summary
 - Added preset for mitochondrial filter settings.
 - Fixed bug where HPO name wasn't displayed in textarea after reloading page.
 - Added possibility to enter OMIM terms in phenotype prioritization filter.
+- Added maximal exon distance field to ``Variants & Effects`` tab.
 
 Full Change List
 ================
@@ -21,6 +22,7 @@ Full Change List
   HPO terms are now also checked for validity in textbox on the fly.
 - Added possibility to enter OMIM terms in phenotype prioritization filter.
   The same textbox as for HPO terms also accepts OMIM terms now.
+- Added maximal exon distance field to ``Variants & Effects`` tab.
 
 -------
 v0.20.0